# Source_code/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: MedSched API
  version: "1.0.0"
  description: |
    API Spec cho hệ thống đặt lịch khám MedSched.

    Tài liệu này là **contract** (bản cam kết) giữa Backend và Frontend:
    - Backend code đúng theo spec này, không tự thêm/bớt field.
    - Frontend dùng spec này để mock data và dựng UI mà không phải chờ Backend.

    Xem README.md cùng thư mục để biết cách dùng (Swagger UI, Postman, mock server).
servers:
  - url: http://localhost:8080
    description: Local development (backend chạy bằng ./gradlew bootRun)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Gửi kèm header: `Authorization: Bearer <accessToken>`.
        Lấy accessToken từ `POST /api/v1/auth/login`.
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
      description: Số thứ tự trang, bắt đầu từ 0.
    SizeParam:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Số bản ghi mỗi trang. Tối đa 100, vượt quá sẽ bị ép về 20.
    SortParam:
      name: sort
      in: query
      required: false
      schema:
        type: string
        example: createdAt,desc
      description: |
        Định dạng `field,direction`. direction nhận `asc` hoặc `desc` (mặc định `asc`).
        Chỉ chấp nhận field trong whitelist: `createdAt`, `updatedAt`, `email`,
        `fullName`, `active`. Field ngoài whitelist -> 400.
        Mặc định khi không truyền: `createdAt,desc`.
    SearchParam:
      name: q
      in: query
      required: false
      schema: { type: string }
      description: Từ khóa tìm kiếm, khớp một phần (LIKE) trên `email` và `fullName`, không phân biệt hoa thường.
  responses:
    Unauthorized:
      description: Chưa đăng nhập, thiếu token, token sai hoặc hết hạn.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example:
            timestamp: "2026-09-14T06:45:47Z"
            status: 401
            error: Unauthorized
            message: Bạn cần đăng nhập để dùng chức năng này
            path: /api/v1/me
            fieldErrors: null
    Forbidden:
      description: Đã đăng nhập nhưng không đủ quyền.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example:
            timestamp: "2026-09-14T06:45:47Z"
            status: 403
            error: Forbidden
            message: Tài khoản của bạn không có quyền thực hiện chức năng này
            path: /api/v1/admin/users
            fieldErrors: null
    NotFound:
      description: Không tìm thấy dữ liệu.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Xung đột dữ liệu (trùng email, trùng mã, slot đã bị giữ...).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ValidationError:
      description: Dữ liệu gửi lên không hợp lệ. Chi tiết từng field nằm trong `fieldErrors`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example:
            timestamp: "2026-09-14T06:45:47Z"
            status: 400
            error: Bad Request
            message: Dữ liệu gửi lên không hợp lệ
            path: /api/v1/auth/register
            fieldErrors:
              email: Email không đúng định dạng
              password: Mật khẩu phải từ 8 đến 72 ký tự
  schemas:
    ApiError:
      type: object
      description: Khuôn lỗi dùng chung cho MỌI API. Frontend chỉ cần viết một hàm xử lý lỗi.
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 400 }
        error: { type: string, example: Bad Request }
        message: { type: string, description: Thông báo tiếng Việt hiển thị được cho người dùng }
        path: { type: string, example: /api/v1/auth/login }
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: { type: string }
          description: Chỉ có khi lỗi validation. Key là tên field, value là thông báo.
    PageResponseUserSummary:
      type: object
      description: Khuôn phân trang chuẩn cho API List.
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/UserSummary' }
        page: { type: integer, example: 0, description: Trang hiện tại (từ 0) }
        size: { type: integer, example: 20, description: Số bản ghi mỗi trang }
        totalItems: { type: integer, format: int64, example: 42, description: Tổng số bản ghi khớp điều kiện }
        totalPages: { type: integer, example: 3 }
...
    UserSummary:
      type: object
      properties:
        userId: { type: string, format: uuid }
        email: { type: string }
        fullName: { type: string }
        phone: { type: string, nullable: true }
        active: { type: boolean }
        roles:
          type: array
          items: { $ref: '#/components/schemas/RoleAssignment' }
          description: Rỗng nghĩa là khách hàng thuần (không có quyền nhân sự).
        createdAt: { type: string, format: date-time }
    RoleAssignment:
      type: object
      properties:
        assignmentId: { type: string, format: uuid, description: Dùng để thu hồi quyền này }
        role: { type: string, enum: [ROLE_DOCTOR, ROLE_STAFF, ROLE_ADMIN] }
        medicalCenterId: { type: string, format: uuid }
        medicalCenterName: { type: string, nullable: true }
        active: { type: boolean }
    CreateStaffRequest:
      type: object
      required: [email, password, fullName, medicalCenterId]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72 }
        fullName: { type: string }
        phone: { type: string, pattern: '^$|^0[0-9]{8,10}$' }
        medicalCenterId: { type: string, format: uuid, description: Chi nhánh mà lễ tân này làm việc }
    CreateDoctorRequest:
      type: object
      required: [email, password, fullName, specialtyId, consultationFee]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72 }
        fullName: { type: string }
        phone: { type: string, pattern: '^$|^0[0-9]{8,10}$' }
        specialtyId: { type: string, format: uuid, description: Chi nhánh được suy ra từ chuyên khoa, không cần truyền riêng }
        academicTitle: { type: string, maxLength: 100 }
        experienceYears: { type: integer, minimum: 0 }
        consultationFee: { type: number, format: double, minimum: 0 }
        roomNumber: { type: string, maxLength: 50 }
    UpdateStatusRequest:
      type: object
      required: [active]
      properties:
        active: { type: boolean, description: false = khóa tài khoản, true = mở lại }
    GrantRoleRequest:
      type: object
      required: [role, medicalCenterId]
      properties:
        role: { type: string, enum: [ROLE_DOCTOR, ROLE_STAFF, ROLE_ADMIN] }
        medicalCenterId: { type: string, format: uuid }
...
paths:
...
  /api/v1/admin/users:
    get:
      tags: [Admin Users]
      summary: Danh sách người dùng (API List chuẩn - pagination, filter, search, sort)
      description: |
        Đây là API List mẫu, áp dụng đủ 4 yêu cầu thiết kế:
        - **Pagination**: `page`, `size` -> response trả `page`, `size`, `totalItems`, `totalPages`
        - **Filter**: `role`, `centerId`
        - **Search**: `q` (khớp một phần trên email và fullName)
        - **Sort**: `sort=field,direction` với whitelist field

        Chỉ `ROLE_ADMIN` gọi được.
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [ALL, CUSTOMER, ROLE_DOCTOR, ROLE_STAFF, ROLE_ADMIN]
            default: ALL
          description: |
            Filter theo vai trò. `CUSTOMER` = tài khoản không có quyền nhân sự nào
            (khách hàng thuần). `ALL` hoặc bỏ trống = lấy tất cả.
        - name: centerId
          in: query
          required: false
          schema: { type: string, format: uuid }
          description: Filter theo chi nhánh. Chỉ có tác dụng khi lọc theo role nhân sự.
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageResponseUserSummary' }
        '400':
          description: Sort field không nằm trong whitelist, hoặc role không hợp lệ.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/users/{userId}:
    get:
      tags: [Admin Users]
      summary: Xem chi tiết một tài khoản
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/users/staff:
    post:
      tags: [Admin Users]
      summary: Tạo tài khoản Lễ tân
      description: Tạo user + cấp quyền ROLE_STAFF tại chi nhánh chỉ định, trong cùng một transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStaffRequest' }
      responses:
        '201':
          description: Tạo thành công.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Không tìm thấy chi nhánh.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/users/doctors:
    post:
      tags: [Admin Users]
      summary: Tạo tài khoản Bác sĩ
      description: |
        Tạo user + cấp quyền ROLE_DOCTOR + tạo hồ sơ bác sĩ (giá khám, phòng, học vị).
        Chi nhánh được suy ra từ `specialtyId` nên không thể gán nhầm bác sĩ vào
        chi nhánh không có chuyên khoa đó.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDoctorRequest' }
      responses:
        '201':
          description: Tạo thành công.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Không tìm thấy chuyên khoa.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/users/{userId}/status:
    patch:
      tags: [Admin Users]
      summary: Khóa / mở khóa tài khoản
      description: Tài khoản bị khóa không đăng nhập được ngay. Admin không tự khóa được chính mình.
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateStatusRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '400':
          description: Tự khóa chính mình.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/users/{userId}/roles:
    post:
      tags: [Admin Users]
      summary: Cấp thêm quyền nhân sự tại một chi nhánh
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GrantRoleRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Tài khoản đã có quyền này tại chi nhánh đó.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /api/v1/admin/users/{userId}/roles/{assignmentId}:
    delete:
      tags: [Admin Users]
      summary: Thu hồi quyền nhân sự
      description: Đánh dấu active = false, KHÔNG xóa dòng, để giữ lịch sử ai từng làm ở đâu.
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: assignmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '400':
          description: Quyền này không thuộc tài khoản đã chọn.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...